+++ illuin_inject/injection_state.py
@@ -2,6 +2,7 @@ from typing import Dict, Optional, TYPE_CHECKING
 
 import attr
 
+from .injector_options import InjectorOptions
 from .provider_registry import ProviderRegistry
 
 if TYPE_CHECKING:
@@ -14,6 +15,7 @@ if TYPE_CHECKING:
 class InjectionState:
     provider_creator: "ProviderCreator"
     binding_registry: "BindingRegistry"
+    options: InjectorOptions = attr.Factory(InjectorOptions)
     parent_state: Optional["InjectionState"] = None
     provider_registry: ProviderRegistry = attr.Factory(ProviderRegistry)
     state_by_module: Dict["AbstractModule", "InjectionState"] = attr.Factory(dict)