+++ illuin_inject/injector_options.py
@@ -1,0 +1,6 @@
+import attr
+
+
+@attr.s(auto_attribs=True)
+class InjectorOptions:
+    auto_bindings: bool = False